requirements.txt
# Core Data Processing
pyspark==3.5.0
delta-spark==3.0.0
pandas==2.1.3
numpy==1.26.2

# Data Quality & Validation
great-expectations==0.18.8
pydantic==2.5.2

# Databricks
databricks-cli==0.18.0
databricks-sdk==0.14.0

# Data Visualization (for local development)
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Excel/File Handling
openpyxl==3.1.2
xlrd==2.0.1

# Database Connectors
sqlalchemy==2.0.23
pyodbc==5.0.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3

# Logging
loguru==0.7.2

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Notebook utilities
ipython==8.18.1
jupyter==1.0.0
nbconvert==7.12.0
